chore: Added utils for lokalise support (#39)

+++ utils/generate_template.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 output_file=message.pot
 if [ -n "$1" ]; then
@@ -6,12 +6,15 @@ if [ -n "$1" ]; then
 fi
 
 if [ "$1" == "--help" ] || [ "$1" == "-h" ]; then
-echo "Usage: prepare_template [FILENAME]
+echo "Usage: generate_template [FILENAME]
 When no [FILENAME] then output is saved to \"message.pot\""
 exit 0
 fi
 
 temp_file=$(mktemp)
-find ../electrum -type f -name "*.py" -o -name "*.kv" > "$temp_file"
+find ../electrum -type f -name "*.py" > "$temp_file"
 xgettext -s --sort-output --no-location --from-code UTF-8 --no-wrap -L Python -f "$temp_file" -o "$output_file"
-rm "$temp_file"
+
+if [ -f "$temp_file" ]; then
+  rm "$temp_file"
+fi
